feat(qos): Moved CloudDriver QOS stuff into orca-qos module
