+++ orca-qos/orca-qos.gradle
@@ -23,4 +23,5 @@ dependencies {
   compile project(":orca-queue")
 
   testCompile project(":orca-test-kotlin")
+  testCompile "com.google.guava:guava-testlib:${spinnaker.version('guava')}"
 }
