+++ orca-qos/src/main/kotlin/com/netflix/spinnaker/orca/qos/retrofit/RequestMetrics.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.orca.reactive
+package com.netflix.spinnaker.orca.qos.retrofit
 
 import java.time.Duration
