+++ orca-qos/src/main/kotlin/com/netflix/spinnaker/orca/qos/retrofit/WindowedRequestMetrics.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.orca.reactive
+package com.netflix.spinnaker.orca.qos.retrofit
 
 import com.google.common.base.Ticker
 import com.google.common.cache.CacheBuilder
