+++ orca-qos/src/test/kotlin/com/netflix/spinnaker/orca/qos/retrofit/WindowedRequestMetricsTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.orca.reactive
+package com.netflix.spinnaker.orca.qos.retrofit
 
 import com.google.common.testing.FakeTicker
 import org.assertj.core.api.Assertions.assertThat
