+++ settings.gradle
@@ -20,7 +20,6 @@ include "orca-extensionpoint",
   "orca-redis",
   "orca-test-redis",
   "orca-retrofit",
-  "orca-retrofit-reactive",
   "orca-front50",
   "orca-bakery",
   "orca-dry-run",
